SQLX support
SQLX compute (P)PSD continuously. If this is available, there is no
point in wasting time computing PPSD with obspy. This is adding a
`PSDs`python object for handling PSD from SQLX in a very similar manner
to the original notebook SeismoSocialDistancing.ipynb.

For now the interface with SQLX is a rather rough system call to SQLX's
script `exPSDhour` via an ssh connection. None of this is optimised or
properly included in the class.

There is also a new very sketchy notebook demonstrating ho to use this
new python object.

It now take a few minutes to get 3-component RMS over several months :)