Add hour map and new module
Following uo on the nice `clockplots` added to the main notebook by
@ThomasLecocq and the growing complexity, I mean to start developing
within a module, than can be then imported in notebooks.

The module is `seismosocialdistancing.py`. I am proposing to put all
our functions there and import the module in all notebooks instead of
creating functions or long codes in notebook cells. In case it is not
obvious, while developing a new function in module, one can reload the
module with:
```
import imp
imp.reload(seismosocialdistancing)
```

I demonstrate how to in sqlxSocialDistancing.ipynb with:
- the sqlx2rms function which produce an object with PSDs and an dRMS
method.
- the hourmap function (poor name?) which makes a map of dRMS aroud the
clock.